refactor: consolidate product update controller and remove unused discount controller
Move product delete controller to dedicated package. Consolidate product update form into details view, removing separate form.html/js files. Remove unused discount controller and field constants with tests. Update log manager to use neat.SortDesc instead of sb.DESC. Fix fakeLogStore count type from int to int64. Add disabled and tabindex attributes to readonly product ID field.

+++ pkg/logadmin/log_manager/log_list_results_test.go
@@ -14,7 +14,7 @@ type fakeLogStore struct {
 	logstore.StoreInterface
 
 	logsToReturn []logstore.LogInterface
-	count        int
+	count        int64
 	listErr      error
 	countErr     error
 }
@@ -30,7 +30,7 @@ func (s *fakeLogStore) LogCount(ctx context.Context, q logstore.LogQueryInterfac
 	if s.countErr != nil {
 		return 0, s.countErr
 	}
-	return int64(s.count), nil
+	return s.count, nil
 }
 
 func TestListLogs_NilApp_ReturnsEmptyNoError(t *testing.T) {
